feat: Separar a instalação de pacotes extras por categorias
- Move packages from `$EXTRA_PKGLIST` to separate categories
- Add choice option for default terminal
- Add output messages for each installation step

The packages currently installed by default by the script are those
that I consider necessary for my use, but not all of them are really
necessary for the system to function.

Users can now decide whether or not to install these packages by
setting the 'INSTALL_*_PKGS' variables to 'true' or 'false' without the
need to manually edit the package lists, although they still can if
they want.