feat: Add the `--needed` flag to the installation of extra packages
Using the `--noconfirm` flag without the `--needed` flag caused already
installed packages to be reinstalled, significantly increasing the
installation time. The installation of video drivers and desktop
environment were not changed to ensure the compatibility of the
packages required by them.

PKGLISTS whose installation was changed to use the `--needed` flag:

- NETWORK_PKGLIST
- TERMINAL_TOOLS_PKGLIST
- FILESYSTEM_PKGLIST
- GENERIC_DRIVERS_PKGLIST
- MULTIMEDIA_PKGLIST
- FONTS_PKGLIST
- EXTRA_PKGLIST